Reinstate original order of Yang nodes
The creation of the dscp-group submodule had the unfortunately effect
of re-ordering the Yang nodes so that the protocol-group node appeared
before the dscp-group node.   This re-ordering is not Yang backwards
compatible.

The commit returns the order of the Ynag node back to their original
ordering by moving the protocol-group out into its own submodule and
including it after the dscp-group submodule.

Fixes: VRVDR-49362