cgnat: do display timeouts value if not in JSON for session detail
Fix for VRVDR-49415 Python traceback with "show cgnat session detail
exclude-inner"

When the command "show cgnat session detail exclude-inner" is executed
the JSON values of "cur_to" and "max_to" are not provided when in
5-tuple mode (and the 3-tuple session was not created by a mapping),
as they are not relevant, as the timeout of the 3-tuple session will
occur when each associated 2-tuple is timeout. So update the script
to not assume they are always provided, and only print the "Timeout:"
line if both values are provided.

(cherry picked from commit 0006f4aaf47e96ea7e1f0d6a9e49acad97e602a3)