cgnat: display timeout as "-" if not in JSON when showing session
Part of fix for VRVDR-49415 Python traceback with "show cgnat session detail
exclude-inner"

The command "show cgnat session exclude-inner" would shown the Timeout
as "None" (due to the python value being 'None') when in 5-tuple
mode, as the timeout "cur_to" is not passed down due to it not being
relevant. This updates it to display "-" instead, which is used for for
other non-relevant values displayed in a table.

(cherry picked from commit 4ac3cbcface5b2c68b77147c075016e116aeaa5a)