Create the resources-group ephemeral VCI component
Create a ephemeral VCI component to take care of creating the necessary
vyatta-dataplane commands to create and delete resources-group objects.

The initial plug-in is for dscp-groups.  Other resources group objects
can be added at a later date.

With dscp-groups now under the control of a VCI component, we can use
VCI component ordering to ensure that resources group objects are
configured before any QoS configuration.

Fixes: VRVDR-49474