cgnat: make per-port timeout node "established" mandatory
Part of fix for VRVDR-49905 CGNAT tcp session-timeout not being set
after session established.

It does not make sense to configure a TCP/UDP port to change the timeout
value for, but then not provide a value. Therefore make the "established"
nodes for TCP/UDP per-port timeouts mandatory, to ensure a value is provided.