cgnat: fix up sending per-port timeout configuration
Part of fix for VRVDR-49905 CGNAT tcp session-timeout not being set
after session established.

The per-port timeout configuration was overwriting the per-protocol-state
timeout configuration, and also only one value configured in a given
commit was being sent to the cstore for use by the dataplane. The code has
been updated to fix this, by ensuring that it writes the different lines
to store to different dictionaries in the temporary "command" dictionary.

Also the code has been updated to delete per-port entries, which
is achieved by setting the timeout value to 0. In the prior code they
would not have been deleted.