cgnat: Add subnet option to nat pool range
VRVDR-50373.  Existing prefix option is changed in the dataplane such
that the first and last addresses are used provided they are not .0
and .255, respectively.   A new subnet option is added, for which the
first and last addresses are never used.

(cherry picked from commit 8dbe8b5fb88c42334be402e1b7ff1d6b58973831)