npf: Change address group show cmd to use python
VRVDR-51350. This builds on the python library file created for inline
display of cgnat policy match address groups.  There is now a single
python function to display address groups for npf, nat pools, and cgnat
policies.  Also simplifies and enhances the api with the data plane, and
the address group show command options.