Add an interface parameter to the npf store_cfg command
VRVDR-46731.  The interface parameter is passed to ctrl.store.  This
allows interface commands for non-existent ephemeral interfaces to be
delayed by the config process until such time that the interface is
created.  This parameter defaults to "ALL", which is the same
default as the 'store' method in the 'Controller' class.