nat: fix issue of python uninit message for exclude
Changes made under vrvdr-50718 involved removing a check for total and
used counts not being in the JSON, but that is required as they do not
appear if the NAT rule has an "exclude" line.

The change here was based on the contents of the vyatta-dp-npf-show-rules
file before it was changed under vrvdr-50718 due to review comments
relating to removing checks for total/used-counts. The code has since
been updated to move the 'total' setting to one place.

VRVDR-51600