SIAD-ACL: Make 'must' path component names qnames
Jira: VRVDR-51619

Within the must statement of a deviation, change the names of
the path component elements to be 'qnames' instead of 'names'.

This makes the names live within an explicit namespace rather
than the default namespace.

It seems that the prior form was valid, and should have worked,
however limitations in our configd currently require the
explicit qname form.