Return empty JSON when resource group state is requested via VCI
When the resource group VCI component is asked for state (get-state) it
does not print anything, so when that is parsed by configd there is the
error "json: error calling MarshalRFC7951 for type ephemera.encodedString:
unexpected end of JSON input" logged. It has been changed to print "{}"
so that it is valid JSON and so prevents this error being logged.

VRVDR-50488