l3acl: Added support for "auto-per-action" to the SW counters
If there are auto-per-action counters in the system configuration, the SW
counters statistics are bundled up by action in order to keep the show
command output consistent between HW and SW counters.

This commit can be reverted when the support for auto-per-action counters
is implemented in the SW path in the dataplane.

VRVDR-51137