Allow 0 as a possible value in a port range for matching pkts
For backward compatibility with 1801, allow 0 as a possible port range
value. Without this, source/destination ports in match rules which contain
0 are failed at validation time. This caused upgrades from 1801 to fail
if they contained 0 in the port range, as in that release they passed
validation. Even though it was intended that 0 could not be used when
configuring, it does actually function as expected when configured in the
dataplane as the code there does support a port range having a 0 value.

The function get_port_type_and_value() needs updated in both FWHelper.pm
and GetPortTypeAndValue.pm. The reason there are two versions, was
validation was being optimised and so the function was moved into its own
perl module to improve performance.  Further changes had not yet been done
for commits, and so they are still using the function in FWHelper.pm,
and also the QoS validation code uses that version, as it has not had
its validation optimised.

Note that SNAT/DNAT have never done the extra validation of the port
values, and so they still do not fail for invalid values checked by the
validation script. However they still need the fixes so that they are
configured with 0-<number> as a port range, rather then a port group.

VRVDR-52668