Add generic packet classifier yang
Add YANG to configure a new generic packet classifier component,
based on the IP packet filter, which allows a set of match rules
to be configured and produce a set of results. Specific features
can then take these results and configure their own actions to
be performed on matching packets.

VRVDR-46091