Add generic packet classifier VCI component
Add a new VCI component to process configuration for a new generic
packet classifier, based on the IP packet filter, which allows a set
of match rules to be configured and produce a set of results. Specific
features can then take these results and configure their own actions to
be performed on matching packets. The GPC VCI component produces a protobuf
configuration describing the rules which can be accessed by features who
wish to use it to configure the dataplane.

VRVDR-46091