Tweak YANG to keep Cisco NCS happy
Part of VRVDR-53065.

The old version of Cisco NCS used by AT&T's Flexware team is unable
to compile vyatta-dtaplane-stats-npf-v1.yang due to the augment of a
node in an RPC added from a grouping.  While it's valid YANG, NCS
can't handle it.  The changes here add a couple of extra grouping
to work around the problem.  Not ideal, but we need to keep the
NCS compilation working.