ippf: ensure cannot match on L4 fields if matching fragments
Add checks to the YANG to ensure that if choosing to match on a
fragment, when you cannot also match on L4 fields like TCP/UDP
ports, and ICMP types and codes. It covers configuring fragmentation
using "fragment any" and also (for IPv6) matching on the header
type being name "ipv6-frag" or number 44. This is done for
both the base and the final protocol for ICMPv4 types/codes, and
just the base protocol for L4 ports.

VRVDR-53316