NPF: Allow for more flexible IPv6 RT RH matches
Jira: VRVDR-53475

The yang only allows two forms of config matching on a RH type field,
either without any explicit protocol match, or with an explicit protocol
match for the routing extension header. It disallows RH type field
matches in combination with any other protocol match.

The dataplane in building its ncode for protocol matching checks to
see if a RH type field is being matched, and if so ignores any explicit
protocol (i.e. RH extension) match. That is because the RH extension
is detected in the packet parse process, and the ncode for the RH type
matching implies that a RH was present; also that an explicit match
would be for a proto-final position, which is never the case with the RH.

So at the moment, it is impossible to configure so as to match a specific
L4 protocol (e.g. UDP) in an extension chain containing a RH.

Hence it would be better if the control plane never sent a protocol match,
neither in the implicit case, nor the explicit case.