GPC: use correct protobuf field names
The vyatta-resources-gpc VCI code had a typo in a protobuf field name.
The VCI code thought that the protobuf field was called 'dst_port'
where as the field is really called 'dest_port'.

Having the wrong field name caused the VCI code to throw an
AttributeError exception that wasn't handled.

Also the VCI code was setting the src_port field if the is_dest
flag was set, and the dest_port field if it wasn't.

Fixes: VRVDR-54132