ippf: fix issue of show cmd matching incorrect rule numbers
When showing ippf rules and filtering on a rule number it was showing
any rule containing a subset of the number entered, e.g. when showing
rule 111 it would also show rule 11 and rule 1.

This was because the rule number entered was converted into a string,
and a check is made of the rule returned back being "in" the string.
The "in" was intended to look at elements of the list, and this happened
for fields other than rule number, as they were already strings. The fix
is to do the conversion to a string of each element in the list of rule
numbers, instead of the whole list. This makes it a list of strings,
and so the "in" check is now correct.

VRVDR-54120