GPC: limit group and result names to 16 characters
The type vy-type:alpha-numeric is based upon the built-in Yang string
type.  By default Yang string types have a maximum length of
18446744073709551615 characters.  As we have to store these names down
in the dataplane, we have to have some sort of sensible limit on these
names.   The value of 16 character has been initially selected as these
names have to be displayed on a 80-column wide terminal.  If longer
names are needed in the future, their length can be extended.

Fixes: VRVDR-54757