J2: Update validation to check not running out of statistics resources
On the J2 there is a maximum number of statistics that is supported
for ACLs for directions "in" and "out", due to the statistic engines
allocated for them.  This adds a check to ensure these maximums are
not exceeded, so configuring too many is detected at configuration
commit time, rather than failing to configure the statistics on
a rule and cause an error log in the journal.

VRVDR-54705