gpc/acl: update error msgs saying 'protocol final' may not be supported
For platforms such as the QAX and J2 you cannot match on a final protocol
('protocol final') and the CLI to do that is deviated away.

However, 2 errors mention 'protocol final' even though it cannot be
set using the config CLI. So as not to confuse a user over the
missing CLI, add "(if supported)" to the error messages than can
be triggered even though 'protocol final' cannot be configured.

VRVDR-54891