dpi: change application type to be a string
With the change from Qosmos to ndpi, application type is now a string
rather than a number, so change the script which shows the rules to
expect a string. This prevents an error message from the script.

Note running this script for ndpi is not available by the end user,
but it is used by the acceptance tests.

VRVDR-55285